# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LamarIn API",
        "description": "REST API Sistem Informasi Lowongan Pekerjaan — Proyek Akhir Kelompok",
        "contact": {
            "email": "kelompok@lamarIn.dev"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Server Utama"
        }
    ],
    "paths": {
...
        "/api/lamaran/{id}": {
            "get": {
                "tags": [
                    "Pelamar"
                ],
                "summary": "Melihat detail satu lamaran (hanya Penyedia pemilik lowongan)",
                "operationId": "1ccf7650853df9f40552d97d15fbf139",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Lamaran",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail lamaran berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Lamaran tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lamaran"
                ],
                "summary": "Batalkan lamaran (hanya jika status masih 'menunggu')",
                "operationId": "1eb13d2c674df5f7610e604f5531d27d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lamaran berhasil dibatalkan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Lamaran tidak ditemukan"
                    },
                    "422": {
                        "description": "Lamaran tidak dapat dibatalkan karena sudah diproses"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/lowongan/{id}/pelamar": {
            "get": {
                "tags": [
                    "Pelamar"
                ],
                "summary": "Melihat semua pelamar pada lowongan milik penyedia",
                "operationId": "2f4f7cc2080f2762e089946d636e9b39",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Lowongan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar pelamar berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden — bukan penyedia atau bukan lowongan milik sendiri"
                    },
                    "404": {
                        "description": "Lowongan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lamaran/{id}/status": {
            "patch": {
                "tags": [
                    "Pelamar"
                ],
                "summary": "Update status lamaran pelamar (hanya Penyedia)",
                "operationId": "9137b3dc7a761fbf76a08a7efda4d35a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Lamaran",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "diproses",
                                        "enum": [
                                            "menunggu",
                                            "diproses",
                                            "wawancara",
                                            "diterima",
                                            "ditolak"
                                        ]
                                    },
                                    "catatan_penyedia": {
                                        "type": "string",
                                        "example": "Jadwal wawancara Senin pukul 09.00",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status lamaran berhasil diperbarui"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden — bukan penyedia"
                    },
                    "404": {
                        "description": "Lamaran tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lamaran/saya": {
            "get": {
                "tags": [
                    "Status Lamaran"
                ],
                "summary": "Melihat semua riwayat lamaran milik pelamar yang login",
                "operationId": "e41a0c2eb5fb3cac1ba78dee51873291",
                "responses": {
                    "200": {
                        "description": "Daftar riwayat lamaran berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden — bukan pelamar"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lamaran/saya/{id}": {
            "get": {
                "tags": [
                    "Status Lamaran"
                ],
                "summary": "Melihat detail satu lamaran milik pelamar berdasarkan ID",
                "operationId": "a2fc1c0074d3620d7cc719ea3d0c1f2f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail lamaran berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden — bukan pelamar"
                    },
                    "404": {
                        "description": "Lamaran tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}